+++ server/src/main/java/soze/multilife/messages/outgoing/PlayerData.java
@@ -4,16 +4,21 @@ import java.util.Map;
 import java.util.Objects;
 
 /**
- * Player data for one {@link Game}.
+ * Player data for one {@link soze.multilife.game.BaseGame}.
  */
 public class PlayerData extends OutgoingMessage {
 
-	public Map<Long, Long> points;
-	public Map<Long, String> names;
-	public Map<Long, String> colors;
-	public Map<Long, String> rules;
+	public Map<Integer, Integer> points;
+	public Map<Integer, String> names;
+	public Map<Integer, String> colors;
+	public Map<Integer, String> rules;
 
-	public PlayerData(Map<Long, Long> points, Map<Long, String> names, Map<Long, String> colors, Map<Long, String> rules) {
+	public PlayerData(
+			Map<Integer, Integer> points,
+			Map<Integer, String> names,
+			Map<Integer, String> colors,
+			Map<Integer, String> rules
+	) {
 		setType(OutgoingType.PLAYER_DATA);
 		this.points = Objects.requireNonNull(points);
 		this.names = Objects.requireNonNull(names);
